Conflicts:
	Frameworks/Lablib/Lablib.xcodeproj/project.pbxproj

+++ Plugins/LLMouseDataDevice/LLMouseDataDevicePIugin.m
@@ -12,7 +12,7 @@
 
 @implementation LLMouseDataDevicePIugin
 
-+ (int)version;
++ (long)version;
 {
 	return kLLPluginVersion;
 }
